+++ src/main/java/database/Application.java
@@ -1,5 +1,7 @@
 package database;
 
+import database.server.DatabaseServer;
+
 public class Application {
 
     public static void main(String[] args) {
