+++ src/main/java/database/server/DatabaseServer.java
@@ -1,4 +1,4 @@
-package database;
+package database.server;
 
 import java.io.IOException;
 import java.net.ServerSocket;
