+++ src/main/java/database/storage/bufferpool/PageId.java
@@ -1,4 +1,4 @@
-package database;
+package database.storage.bufferpool;
 
 public record PageId(String fileName, int pageNumber) {
 }
