+++ src/main/java/database/storage/bufferpool/strategy/LRUStrategy.java
@@ -1,4 +1,4 @@
-package database;
+package database.storage.bufferpool.strategy;
 
 import java.util.Collection;
 import java.util.LinkedHashSet;
