+++ src/main/java/database/storage/bufferpool/strategy/PageReplacementStrategy.java
@@ -1,4 +1,4 @@
-package database;
+package database.storage.bufferpool.strategy;
 
 import java.util.Collection;
