+++ src/main/java/database/storage/page/Record.java
@@ -1,4 +1,4 @@
-package database.engine.page;
+package database.storage.page;
 
 import java.nio.ByteBuffer;
