+++ src/test/java/database/storage/bufferpool/strategy/LRUStrategyTest.java
@@ -1,4 +1,4 @@
-package database;
+package database.storage.bufferpool.strategy;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
