+++ src/test/java/database/storage/page/PageTypeTest.java
@@ -1,4 +1,4 @@
-package database.engine.page;
+package database.storage.page;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertThrows;
